requirements.txt
altgraph==0.17
asn1crypto==0.22.0
astroid==2.3.3
attrs==15.2.0
Automat==0.0.0
certifi==2016.2.28
cffi==1.10.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
constantly==15.1.0
cryptography==1.8.1
cssselect==1.1.0
DBUtils==1.3
defusedxml==0.6.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Login==0.5.0
future==0.18.2
hyperlink==17.1.1
idna==2.6
incremental==16.10.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
json5==0.9.4
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.13.1
openpyxl==3.0.3
packaging==16.8
parsel==1.2.0
pefile==2019.4.18
pg8000==1.13.2
Pillow==7.1.1
psycopg2==2.8.4
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.18
pycurl==7.43.0.5
PyDispatcher==2.0.5
PyInstaller==3.6
pylint==2.4.4
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==17.0.0
pyparsing==2.2.0
pyquery==1.4.1
pyspider==0.3.10
pywin32==220
pywin32-ctypes==0.2.0
PyYAML==5.3
queuelib==1.4.2
redis==3.4.1
requests==2.14.2
scramp==1.1.0
Scrapy==1.3.3
service-identity==17.0.0
six==1.10.0
tblib==1.6.0
tornado==4.5.3
Twisted==17.5.0
typed-ast==1.4.1
u-msgpack-python==2.5.2
w3lib==1.17.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.11.2
WsgiDAV==3.0.3
zope.interface==4.4.2
